refactor: convert to TypeScript with openapi-types
- Add tsconfig.json with strict mode
- Add tsx, typescript, openapi-types, and @types packages
- Create scripts/lib/openapi.ts with generic OpenAPI utilities
- Convert combine-bigcommerce-catalog.mjs to TypeScript
- Add tsx shebang for direct execution
- Update biome config to include .ts files
- Add typecheck script to package.json
