+++ package.json
@@ -1,13 +1,14 @@
 {
   "name": "openapi-specs",
   "version": "0.0.1",
+  "type": "module",
   "description": "",
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "spectral lint ./specs/**/*.{json,yml,yaml}",
     "lint:biome": "biome check",
-    "format": "biome check --fix"
+    "format": "biome check --fix",
+    "typecheck": "tsc --noEmit"
   },
   "repository": {
     "type": "git",
@@ -20,14 +21,20 @@
   },
   "homepage": "https://github.com/takeshape/openapi-specs#readme",
   "devDependencies": {
-    "@biomejs/biome": "2.2.4",
     "@apidevtools/swagger-parser": "^12.1.0",
+    "@biomejs/biome": "2.2.4",
     "@stoplight/spectral-cli": "^6.10.1",
     "@takeshape/util": "^8.271.1",
+    "@types/js-yaml": "^4.0.9",
+    "@types/lodash": "^4.17.16",
+    "@types/node": "^22.15.21",
     "cheerio": "^1.0.0-rc.12",
     "got": "^12.5.3",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
-    "p-map": "^5.5.0"
+    "openapi-types": "^12.1.3",
+    "p-map": "^5.5.0",
+    "tsx": "^4.19.4",
+    "typescript": "^5.8.3"
   }
 }
